refactor(friends): 친구 페이지 분할
