+++ components/community/Friends/FriendsRightContents/FriendStock/StockList/index.tsx
