+++ components/community/Friends/FriendsRightContents/FriendStock/StockList/style.ts
