+++ components/community/Friends/FriendsRightContents/FriendStock/index.tsx
