+++ components/community/Friends/FriendsRightContents/ReqFriends/WaitingFriends/EachWaitingFriend.tsx
