+++ components/community/Friends/FriendsRightContents/ReqFriends/WaitingFriends/index.tsx
