+++ components/community/Friends/FriendsRightContents/SearchFriend/SearchResults/ResultsList/style.ts
