+++ components/community/Friends/FriendsRightContents/SearchFriend/SearchResults/index.tsx
