+++ components/community/Friends/FriendsRightContents/SearchFriend/style.ts
