+++ components/community/Friends/index.tsx
@@ -1,21 +1,29 @@
 import styled from 'styled-components';
 import Profiles from './Profiles';
-import SearchFriend from './SearchFriend';
 import { useState } from 'react';
+import Card from 'components/common/Card';
 
 const CommunitySection = styled.section`
   display: flex;
   gap: 7.1rem;
 `;
 
+interface IFriendsProps {
+  children?: React.ReactNode;
+}
+
 /** community - 친구 */
-const Friends = () => {
-  const [reqFriends, setReqFriends] = useState(false);
+const Friends = ({ children }: IFriendsProps) => {
+  const [isFriendStock, setIsFriendStock] = useState(false);
+  const [isReqFriends, setIsReqFriends] = useState(false);
+  const [isSearchFriend, setIsSearchFriend] = useState(false);
 
   return (
     <CommunitySection>
-      <Profiles setReqFriends={setReqFriends} />
-      <SearchFriend reqFriends={reqFriends} setReqFriends={setReqFriends} />
+      <Profiles setIsReqFriends={setIsReqFriends} />
+      <Card width="65.9rem" height="73.8rem">
+        {children}
+      </Card>
     </CommunitySection>
   );
 };
